improve layout

+++ src/Todo.jsx
@@ -1,34 +1,14 @@
-import {
-  Button,
-  List,
-  ListItem,
-  ListItemAvatar,
-  ListItemText,
-  Modal,
-} from "@mui/material";
 import { deleteDoc, doc, updateDoc } from "firebase/firestore";
 import { db } from "./firebase";
-import DeleteForeverIcon from "@mui/icons-material/DeleteForever";
 import { useState } from "react";
-import Box from "@mui/material/Box";
+import "./Todo.css";
 
-const style = {
-  position: "absolute",
-  top: "50%",
-  left: "50%",
-  transform: "translate(-50%, -50%)",
-  width: 400,
-  bgcolor: "background.paper",
-  border: "2px solid #000",
-  boxShadow: 24,
-  p: 4,
-};
-
-function Todo({ text, id }) {
+function Todo({ text, id, time }) {
   const [open, setOpen] = useState(false);
-  const [input, setInput] = useState('');
+  const [input, setInput] = useState("");
 
   const handleDelete = () => {
+    if (window.confirm("Are you sure you want to delete this todo?"))
     deleteDoc(doc(db, "todos", id));
   };
 
@@ -48,32 +28,16 @@ function Todo({ text, id }) {
 
   return (
     <>
-      <Modal
-        open={open}
-        onClose={handleClose}
-        aria-labelledby="modal-modal-title"
-        aria-describedby="modal-modal-description"
-      >
-        <Box sx={style}>
-          <div>
-            <h1>Update Todo</h1>
-            <input placeholder={text} value={input} onChange={(e) => setInput(e.target.value)} />
-            <Button onClick={updateTodo}>Update Todo</Button>
-          </div>
-        </Box>
-      </Modal>
-      <List className="todo_list">
-        <ListItem>
-          <ListItemAvatar></ListItemAvatar>
-          <ListItemText primary={text} secondary="Dummy Deadline" />
-        </ListItem>
-        <button onClick={handleOpen}>Edit</button>
-        <DeleteForeverIcon
-          onClick={handleDelete}
-          variant="contained"
-          color="primary"
-        />
-      </List>
+      <div className="todo">
+        <div className="container1">
+          <h3>{text}</h3>
+          <p>{time}</p>
+        </div>
+        <div className="container2">
+          <button className="edit" onClick={handleOpen} >Edit</button>
+          <button className="delete" onClick={handleDelete}>Delete</button>
+        </div>
+      </div>
     </>
   );
 }
